chore: SRS-918 surface operation names in Graphql requests
