+++ cats-frontend/src/index.tsx
@@ -1,12 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import {
-  ApolloClient,
-  InMemoryCache,
-  ApolloProvider,
-  createHttpLink,
-} from '@apollo/client';
-import { setContext } from '@apollo/client/link/context';
+import { ApolloProvider } from '@apollo/client';
 
 import './index.css';
 import reportWebVitals from './reportWebVitals';
@@ -18,32 +12,11 @@ import { UserManagerSettings } from 'oidc-client-ts';
 import { getClientSettings } from './app/auth/UserManagerSetting';
 import { RouterProvider } from 'react-router-dom';
 import siteRouter from './app/routes/Routes';
-import { getUser } from './app/helpers/utility';
-import { API, GRAPHQL } from './app/helpers/endpoints';
+import { apolloClientInstance } from './apollo';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLDivElement,
 );
-
-const httpLink = createHttpLink({
-  uri: `${API}${GRAPHQL}`,
-});
-
-const authLink = setContext((_, { headers }) => {
-  const user = getUser();
-
-  return {
-    headers: {
-      ...headers,
-      authorization: user?.access_token ? `Bearer ${user.access_token}` : '',
-    },
-  };
-});
-
-const client = new ApolloClient({
-  link: authLink.concat(httpLink),
-  cache: new InMemoryCache(),
-});
 
 // This is a required callback for proper auth experience. From the `react-oidc-context` docs:
 // "You must provide an implementation of onSigninCallback to oidcConfig to remove the payload from the URL upon successful login.
@@ -56,7 +29,7 @@ const authOptions: UserManagerSettings = getClientSettings();
 root.render(
   <React.StrictMode>
     <AuthProvider {...authOptions} onSigninCallback={onSigninCallback}>
-      <ApolloProvider client={client}>
+      <ApolloProvider client={apolloClientInstance}>
         <Provider store={store}>
           <RouterProvider router={siteRouter} />
         </Provider>
